# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReytingOLAP API
  version: 1.0.0
  description: |
    Talabalar reytingini OLAP tahlili asosida monitoring qilish AT.
    SRAP-2026 — REST API spetsifikatsiyasi (TZ 4.2.1, 8.1).
  contact:
    name: Loyiha rahbari
servers:
  - url: https://api.reytingolap.uz/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /bot/link-parent:
    post:
      tags: [bot]
      summary: Ota-onaning talabaga bog'lanish so'rovi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [telegram_id, talaba_hemis_id]
              properties:
                telegram_id:      { type: integer, format: int64 }
                talaba_hemis_id:  { type: string }
      responses:
        '202': { description: Tasdiqlash so'rovi talabaga yuborildi }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...